requirements.txt
aiohttp
qrcode[pil]
Pillow